backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
nylas>=6.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
